separate rpc/explorer client
