+++ .gitignore
@@ -82,6 +82,7 @@ _cgo_gotypes.go
 _cgo_export.*
 
 _testmain.go
+__debug_bin
 
 *.exe
 *.test
