+++ platform/filecoin/base.go
@@ -2,16 +2,20 @@ package filecoin
 
 import (
 	"github.com/trustwallet/blockatlas/pkg/blockatlas"
+	"github.com/trustwallet/blockatlas/platform/filecoin/explorer"
+	"github.com/trustwallet/blockatlas/platform/filecoin/rpc"
 	"github.com/trustwallet/golibs/coin"
 )
 
 type Platform struct {
-	client Client
+	client   rpc.Client
+	explorer explorer.Client
 }
 
-func Init(api string) *Platform {
+func Init(api, explorerApi string) *Platform {
 	p := &Platform{
-		client: Client{blockatlas.InitClient(api)},
+		client:   rpc.Client{Request: blockatlas.InitClient(api)},
+		explorer: explorer.Client{Request: blockatlas.InitClient(explorerApi)},
 	}
 	return p
 }
